Update CreatePost.jsx
First thing implemented is a useState hook declaring files along with its setter function setFiles.  It is implemnted with the onChange events.

Second thing I did was createNewPost function.
In this function:

A new formData object named data is created.
The title, summary and content values are added to the form data.

I then take the first file from the files array and add it to form data using data.set('file', files[0].

I then fetch the function to a send POST request with the form data to the endpoint that I am now going to implement in my api.