Update HomePage.jsx
In this update:

I am using the useState and useEffect hooks to fetch and display posts from the server.   These hooks manage the state of the posts and trigger the data fetching process.

At the end I use conditonal rendering check if there are posts in the 'posts' array.  This will map over the array and renders individual 'Post' components , spreading the properties of each post as props.